Add Propagator for Chain Sync
