+++ ethcore/sync/src/chain/mod.rs
@@ -88,6 +88,7 @@
 /// All other messages are ignored.
 ///
 
+mod propagator;
 mod requester;
 mod supplier;
 
@@ -101,13 +102,12 @@ use ethereum_types::{H256, U256};
 use plain_hasher::H256FastMap;
 use parking_lot::RwLock;
 use bytes::Bytes;
-use rlp::{Rlp, RlpStream, DecoderError, Encodable};
+use rlp::{Rlp, RlpStream, DecoderError};
 use network::{self, PeerId, PacketId};
 use ethcore::header::{BlockNumber, Header as BlockHeader};
 use ethcore::client::{BlockChainClient, BlockStatus, BlockId, BlockChainInfo, BlockImportError, BlockImportErrorKind, BlockQueueInfo};
 use ethcore::error::*;
 use ethcore::snapshot::{ManifestData, RestorationStatus};
-use transaction::SignedTransaction;
 use sync_io::SyncIo;
 use super::{WarpSync, SyncConfig};
 use block_sync::{BlockDownloader, BlockDownloaderImportError as DownloaderImportError, DownloadAction};
@@ -118,8 +118,9 @@ use private_tx::PrivateTxHandler;
 use transactions_stats::{TransactionsStats, Stats as TransactionStats};
 use transaction::UnverifiedTransaction;
 
-use self::supplier::SyncSupplier;
+use self::propagator::SyncPropagator;
 use self::requester::SyncRequester;
+use self::supplier::SyncSupplier;
 
 known_heap_size!(0, PeerInfo);
 
@@ -358,12 +359,12 @@ impl PeerInfo {
 }
 
 #[cfg(not(test))]
-mod random {
+pub mod random {
 	use rand;
 	pub fn new() -> rand::ThreadRng { rand::thread_rng() }
 }
 #[cfg(test)]
-mod random {
+pub mod random {
 	use rand::{self, SeedableRng};
 	pub fn new() -> rand::XorShiftRng { rand::XorShiftRng::from_seed([0, 1, 2, 3]) }
 }
@@ -1387,14 +1388,6 @@ impl ChainSync {
 		false
 	}
 
-	/// Generic packet sender
-	fn send_packet(&mut self, sync: &mut SyncIo, peer_id: PeerId, packet_id: PacketId, packet: Bytes) {
-		if let Err(e) = sync.send(peer_id, packet_id, packet) {
-			debug!(target:"sync", "Error sending packet: {:?}", e);
-			sync.disconnect_peer(peer_id);
-		}
-	}
-
 	/// Called when peer sends us new transactions
 	fn on_peer_transactions(&mut self, io: &mut SyncIo, peer_id: PeerId, r: &Rlp) -> Result<(), PacketDecodeError> {
 		// Accept transactions only when fully synced
@@ -1632,236 +1625,6 @@ impl ChainSync {
 		self.peers.iter().filter_map(|(id, p)| if p.protocol_version >= PAR_PROTOCOL_VERSION_3 { Some(*id) } else { None }).collect()
 	}
 
-	/// propagates latest block to a set of peers
-	fn propagate_blocks(&mut self, chain_info: &BlockChainInfo, io: &mut SyncIo, blocks: &[H256], peers: &[PeerId]) -> usize {
-		trace!(target: "sync", "Sending NewBlocks to {:?}", peers);
-		let mut sent = 0;
-		for peer_id in peers {
-			if blocks.is_empty() {
-				let rlp =  ChainSync::create_latest_block_rlp(io.chain());
-				self.send_packet(io, *peer_id, NEW_BLOCK_PACKET, rlp);
-			} else {
-				for h in blocks {
-					let rlp =  ChainSync::create_new_block_rlp(io.chain(), h);
-					self.send_packet(io, *peer_id, NEW_BLOCK_PACKET, rlp);
-				}
-			}
-			if let Some(ref mut peer) = self.peers.get_mut(peer_id) {
-				peer.latest_hash = chain_info.best_block_hash.clone();
-			}
-			sent += 1;
-		}
-		sent
-	}
-
-	/// propagates new known hashes to all peers
-	fn propagate_new_hashes(&mut self, chain_info: &BlockChainInfo, io: &mut SyncIo, peers: &[PeerId]) -> usize {
-		trace!(target: "sync", "Sending NewHashes to {:?}", peers);
-		let mut sent = 0;
-		let last_parent = *io.chain().best_block_header().parent_hash();
-		for peer_id in peers {
-			sent += match ChainSync::create_new_hashes_rlp(io.chain(), &last_parent, &chain_info.best_block_hash) {
-				Some(rlp) => {
-					{
-						if let Some(ref mut peer) = self.peers.get_mut(peer_id) {
-							peer.latest_hash = chain_info.best_block_hash.clone();
-						}
-					}
-					self.send_packet(io, *peer_id, NEW_BLOCK_HASHES_PACKET, rlp);
-					1
-				},
-				None => 0
-			}
-		}
-		sent
-	}
-
-	/// propagates new transactions to all peers
-	pub fn propagate_new_transactions(&mut self, io: &mut SyncIo) -> usize {
-		// Early out if nobody to send to.
-		if self.peers.is_empty() {
-			return 0;
-		}
-
-		let transactions = io.chain().ready_transactions();
-		if transactions.is_empty() {
-			return 0;
-		}
-
-		let (transactions, service_transactions): (Vec<_>, Vec<_>) = transactions.iter()
-			.map(|tx| tx.signed())
-			.partition(|tx| !tx.gas_price.is_zero());
-
-		// usual transactions could be propagated to all peers
-		let mut affected_peers = HashSet::new();
-		if !transactions.is_empty() {
-			let peers = self.select_peers_for_transactions(|_| true);
-			affected_peers = self.propagate_transactions_to_peers(io, peers, transactions);
-		}
-
-		// most of times service_transactions will be empty
-		// => there's no need to merge packets
-		if !service_transactions.is_empty() {
-			let service_transactions_peers = self.select_peers_for_transactions(|peer_id| accepts_service_transaction(&io.peer_info(*peer_id)));
-			let service_transactions_affected_peers = self.propagate_transactions_to_peers(io, service_transactions_peers, service_transactions);
-			affected_peers.extend(&service_transactions_affected_peers);
-		}
-
-		affected_peers.len()
-	}
-
-	fn select_peers_for_transactions<F>(&self, filter: F) -> Vec<PeerId>
-		where F: Fn(&PeerId) -> bool {
-		// sqrt(x)/x scaled to max u32
-		let fraction = ((self.peers.len() as f64).powf(-0.5) * (u32::max_value() as f64).round()) as u32;
-		let small = self.peers.len() < MIN_PEERS_PROPAGATION;
-
-		let mut random = random::new();
-		self.peers.keys()
-			.cloned()
-			.filter(filter)
-			.filter(|_| small || random.next_u32() < fraction)
-			.take(MAX_PEERS_PROPAGATION)
-			.collect()
-	}
-
-	fn propagate_transactions_to_peers(&mut self, io: &mut SyncIo, peers: Vec<PeerId>, transactions: Vec<&SignedTransaction>) -> HashSet<PeerId> {
-		let all_transactions_hashes = transactions.iter()
-			.map(|tx| tx.hash())
-			.collect::<HashSet<H256>>();
-		let all_transactions_rlp = {
-			let mut packet = RlpStream::new_list(transactions.len());
-			for tx in &transactions { packet.append(&**tx); }
-			packet.out()
-		};
-
-		// Clear old transactions from stats
-		self.transactions_stats.retain(&all_transactions_hashes);
-
-		// sqrt(x)/x scaled to max u32
-		let block_number = io.chain().chain_info().best_block_number;
-
-		let lucky_peers = {
-			peers.into_iter()
-				.filter_map(|peer_id| {
-					let stats = &mut self.transactions_stats;
-					let peer_info = self.peers.get_mut(&peer_id)
-						.expect("peer_id is form peers; peers is result of select_peers_for_transactions; select_peers_for_transactions selects peers from self.peers; qed");
-
-					// Send all transactions
-					if peer_info.last_sent_transactions.is_empty() {
-						// update stats
-						for hash in &all_transactions_hashes {
-							let id = io.peer_session_info(peer_id).and_then(|info| info.id);
-							stats.propagated(hash, id, block_number);
-						}
-						peer_info.last_sent_transactions = all_transactions_hashes.clone();
-						return Some((peer_id, all_transactions_hashes.len(), all_transactions_rlp.clone()));
-					}
-
-					// Get hashes of all transactions to send to this peer
-					let to_send = all_transactions_hashes.difference(&peer_info.last_sent_transactions)
-						.take(MAX_TRANSACTIONS_TO_PROPAGATE)
-						.cloned()
-						.collect::<HashSet<_>>();
-					if to_send.is_empty() {
-						return None;
-					}
-
-					// Construct RLP
-					let (packet, to_send) = {
-						let mut to_send = to_send;
-						let mut packet = RlpStream::new();
-						packet.begin_unbounded_list();
-						let mut pushed = 0;
-						for tx in &transactions {
-							let hash = tx.hash();
-							if to_send.contains(&hash) {
-								let mut transaction = RlpStream::new();
-								tx.rlp_append(&mut transaction);
-								let appended = packet.append_raw_checked(&transaction.drain(), 1, MAX_TRANSACTION_PACKET_SIZE);
-								if !appended {
-									// Maximal packet size reached just proceed with sending
-									debug!("Transaction packet size limit reached. Sending incomplete set of {}/{} transactions.", pushed, to_send.len());
-									to_send = to_send.into_iter().take(pushed).collect();
-									break;
-								}
-								pushed += 1;
-							}
-						}
-						packet.complete_unbounded_list();
-						(packet, to_send)
-					};
-
-					// Update stats
-					let id = io.peer_session_info(peer_id).and_then(|info| info.id);
-					for hash in &to_send {
-						// update stats
-						stats.propagated(hash, id, block_number);
-					}
-
-					peer_info.last_sent_transactions = all_transactions_hashes
-						.intersection(&peer_info.last_sent_transactions)
-						.chain(&to_send)
-						.cloned()
-						.collect();
-					Some((peer_id, to_send.len(), packet.out()))
-				})
-				.collect::<Vec<_>>()
-		};
-
-		// Send RLPs
-		let mut peers = HashSet::new();
-		if lucky_peers.len() > 0 {
-			let mut max_sent = 0;
-			let lucky_peers_len = lucky_peers.len();
-			for (peer_id, sent, rlp) in lucky_peers {
-				peers.insert(peer_id);
-				self.send_packet(io, peer_id, TRANSACTIONS_PACKET, rlp);
-				trace!(target: "sync", "{:02} <- Transactions ({} entries)", peer_id, sent);
-				max_sent = cmp::max(max_sent, sent);
-			}
-			debug!(target: "sync", "Sent up to {} transactions to {} peers.", max_sent, lucky_peers_len);
-		}
-
-		peers
-	}
-
-	fn propagate_latest_blocks(&mut self, io: &mut SyncIo, sealed: &[H256]) {
-		let chain_info = io.chain().chain_info();
-		if (((chain_info.best_block_number as i64) - (self.last_sent_block_number as i64)).abs() as BlockNumber) < MAX_PEER_LAG_PROPAGATION {
-			let mut peers = self.get_lagging_peers(&chain_info);
-			if sealed.is_empty() {
-				let hashes = self.propagate_new_hashes(&chain_info, io, &peers);
-				peers = ChainSync::select_random_peers(&peers);
-				let blocks = self.propagate_blocks(&chain_info, io, sealed, &peers);
-				if blocks != 0 || hashes != 0 {
-					trace!(target: "sync", "Sent latest {} blocks and {} hashes to peers.", blocks, hashes);
-				}
-			} else {
-				self.propagate_blocks(&chain_info, io, sealed, &peers);
-				self.propagate_new_hashes(&chain_info, io, &peers);
-				trace!(target: "sync", "Sent sealed block to all peers");
-			};
-		}
-		self.last_sent_block_number = chain_info.best_block_number;
-	}
-
-	/// Distribute valid proposed blocks to subset of current peers.
-	fn propagate_proposed_blocks(&mut self, io: &mut SyncIo, proposed: &[Bytes]) {
-		let peers = self.get_consensus_peers();
-		trace!(target: "sync", "Sending proposed blocks to {:?}", peers);
-		for block in proposed {
-			let rlp = ChainSync::create_block_rlp(
-				block,
-				io.chain().chain_info().total_difficulty
-			);
-			for peer_id in &peers {
-				self.send_packet(io, *peer_id, NEW_BLOCK_PACKET, rlp.clone());
-			}
-		}
-	}
-
 	/// Maintain other peers. Send out any new blocks and transactions
 	pub fn maintain_sync(&mut self, io: &mut SyncIo) {
 		self.maybe_start_snapshot_sync(io);
@@ -1875,8 +1638,8 @@ impl ChainSync {
 
 		if !is_syncing || !sealed.is_empty() || !proposed.is_empty() {
 			trace!(target: "sync", "Propagating blocks, state={:?}", self.state);
-			self.propagate_latest_blocks(io, sealed);
-			self.propagate_proposed_blocks(io, proposed);
+			SyncPropagator::propagate_latest_blocks(self, io, sealed);
+			SyncPropagator::propagate_proposed_blocks(self, io, proposed);
 		}
 		if !invalid.is_empty() {
 			trace!(target: "sync", "Bad blocks in the queue, restarting");
@@ -1900,15 +1663,6 @@ impl ChainSync {
 		Ok(())
 	}
 
-	/// Broadcast consensus message to peers.
-	pub fn propagate_consensus_packet(&mut self, io: &mut SyncIo, packet: Bytes) {
-		let lucky_peers = ChainSync::select_random_peers(&self.get_consensus_peers());
-		trace!(target: "sync", "Sending consensus packet to {:?}", lucky_peers);
-		for peer_id in lucky_peers {
-			self.send_packet(io, peer_id, CONSENSUS_DATA_PACKET, packet.clone());
-		}
-	}
-
 	/// Called when peer sends us new private transaction packet
 	fn on_private_transaction(&self, _io: &mut SyncIo, peer_id: PeerId, r: &Rlp) -> Result<(), PacketDecodeError> {
 		if !self.peers.get(&peer_id).map_or(false, |p| p.can_sync()) {
@@ -1923,16 +1677,6 @@ impl ChainSync {
 		}
 		Ok(())
 	}
-
-	/// Broadcast private transaction message to peers.
-	pub fn propagate_private_transaction(&mut self, io: &mut SyncIo, packet: Bytes) {
-		let lucky_peers = ChainSync::select_random_peers(&self.get_private_transaction_peers());
-		trace!(target: "sync", "Sending private transaction packet to {:?}", lucky_peers);
-		for peer_id in lucky_peers {
-			self.send_packet(io, peer_id, PRIVATE_TRANSACTION_PACKET, packet.clone());
-		}
-	}
-
 	/// Called when peer sends us signed private transaction packet
 	fn on_signed_private_transaction(&self, _io: &mut SyncIo, peer_id: PeerId, r: &Rlp) -> Result<(), PacketDecodeError> {
 		if !self.peers.get(&peer_id).map_or(false, |p| p.can_sync()) {
@@ -1947,32 +1691,25 @@ impl ChainSync {
 		Ok(())
 	}
 
+	/// propagates new transactions to all peers
+	pub fn propagate_new_transactions(&mut self, io: &mut SyncIo) -> usize {
+		SyncPropagator::propagate_new_transactions(self, io)
+	}
+
+	/// Broadcast consensus message to peers.
+	pub fn propagate_consensus_packet(&mut self, io: &mut SyncIo, packet: Bytes) {
+		SyncPropagator::propagate_consensus_packet(self, io, packet);
+	}
+
+	/// Broadcast private transaction message to peers.
+	pub fn propagate_private_transaction(&mut self, io: &mut SyncIo, packet: Bytes) {
+		SyncPropagator::propagate_private_transaction(self, io, packet);
+	}
+
 	/// Broadcast signed private transaction message to peers.
 	pub fn propagate_signed_private_transaction(&mut self, io: &mut SyncIo, packet: Bytes) {
-		let lucky_peers = ChainSync::select_random_peers(&self.get_private_transaction_peers());
-		trace!(target: "sync", "Sending signed private transaction packet to {:?}", lucky_peers);
-		for peer_id in lucky_peers {
-			self.send_packet(io, peer_id, SIGNED_PRIVATE_TRANSACTION_PACKET, packet.clone());
-		}
+		SyncPropagator::propagate_signed_private_transaction(self, io, packet);
 	}
-
-}
-
-/// Checks if peer is able to process service transactions
-fn accepts_service_transaction(client_id: &str) -> bool {
-	// Parity versions starting from this will accept service-transactions
-	const SERVICE_TRANSACTIONS_VERSION: (u32, u32) = (1u32, 6u32);
-	// Parity client string prefix
-	const PARITY_CLIENT_ID_PREFIX: &'static str = "Parity/v";
-
-	if !client_id.starts_with(PARITY_CLIENT_ID_PREFIX) {
-		return false;
-	}
-	let ver: Vec<u32> = client_id[PARITY_CLIENT_ID_PREFIX.len()..].split('.')
-		.take(2)
-		.filter_map(|s| s.parse().ok())
-		.collect();
-	ver.len() == 2 && (ver[0] > SERVICE_TRANSACTIONS_VERSION.0 || (ver[0] == SERVICE_TRANSACTIONS_VERSION.0 && ver[1] >= SERVICE_TRANSACTIONS_VERSION.1))
 }
 
 #[cfg(test)]
@@ -2144,7 +1881,7 @@ pub mod tests {
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
 
 		let peers = sync.get_lagging_peers(&chain_info);
-		let peer_count = sync.propagate_new_hashes(&chain_info, &mut io, &peers);
+		let peer_count = SyncPropagator::propagate_new_hashes(&mut sync, &chain_info, &mut io, &peers);
 
 		// 1 message should be send
 		assert_eq!(1, io.packets.len());
@@ -2164,7 +1901,7 @@ pub mod tests {
 		let ss = TestSnapshotService::new();
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
 		let peers = sync.get_lagging_peers(&chain_info);
-		let peer_count = sync.propagate_blocks(&chain_info, &mut io, &[], &peers);
+		let peer_count = SyncPropagator::propagate_blocks(&mut sync, &chain_info, &mut io, &[], &peers);
 
 		// 1 message should be send
 		assert_eq!(1, io.packets.len());
@@ -2185,7 +1922,7 @@ pub mod tests {
 		let ss = TestSnapshotService::new();
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
 		let peers = sync.get_lagging_peers(&chain_info);
-		let peer_count = sync.propagate_blocks(&chain_info, &mut io, &[hash.clone()], &peers);
+		let peer_count = SyncPropagator::propagate_blocks(&mut sync ,&chain_info, &mut io, &[hash.clone()], &peers);
 
 		// 1 message should be send
 		assert_eq!(1, io.packets.len());
@@ -2224,7 +1961,7 @@ pub mod tests {
 			});
 		let ss = TestSnapshotService::new();
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		sync.propagate_proposed_blocks(&mut io, &[block]);
+		SyncPropagator::propagate_proposed_blocks(&mut sync, &mut io, &[block]);
 
 		// 1 message should be sent
 		assert_eq!(1, io.packets.len());
@@ -2241,13 +1978,13 @@ pub mod tests {
 		let queue = RwLock::new(VecDeque::new());
 		let ss = TestSnapshotService::new();
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		let peer_count = sync.propagate_new_transactions(&mut io);
+		let peer_count = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 		// Try to propagate same transactions for the second time
-		let peer_count2 = sync.propagate_new_transactions(&mut io);
+		let peer_count2 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 		// Even after new block transactions should not be propagated twice
 		sync.chain_new_blocks(&mut io, &[], &[], &[], &[], &[], &[]);
 		// Try to propagate same transactions for the third time
-		let peer_count3 = sync.propagate_new_transactions(&mut io);
+		let peer_count3 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 
 		// 1 message should be send
 		assert_eq!(1, io.packets.len());
@@ -2268,7 +2005,7 @@ pub mod tests {
 		let queue = RwLock::new(VecDeque::new());
 		let ss = TestSnapshotService::new();
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		let peer_count = sync.propagate_new_transactions(&mut io);
+		let peer_count = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 		io.chain.insert_transaction_to_queue();
 		// New block import should not trigger propagation.
 		// (we only propagate on timeout)
@@ -2292,10 +2029,10 @@ pub mod tests {
 		let queue = RwLock::new(VecDeque::new());
 		let ss = TestSnapshotService::new();
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		let peer_count = sync.propagate_new_transactions(&mut io);
+		let peer_count = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 		sync.chain_new_blocks(&mut io, &[], &[], &[], &[], &[], &[]);
 		// Try to propagate same transactions for the second time
-		let peer_count2 = sync.propagate_new_transactions(&mut io);
+		let peer_count2 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 
 		assert_eq!(0, io.packets.len());
 		assert_eq!(0, peer_count);
@@ -2313,7 +2050,7 @@ pub mod tests {
 		// should sent some
 		{
 			let mut io = TestIo::new(&mut client, &ss, &queue, None);
-			let peer_count = sync.propagate_new_transactions(&mut io);
+			let peer_count = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 			assert_eq!(1, io.packets.len());
 			assert_eq!(1, peer_count);
 		}
@@ -2322,9 +2059,9 @@ pub mod tests {
 		let (peer_count2, peer_count3) = {
 			let mut io = TestIo::new(&mut client, &ss, &queue, None);
 			// Propagate new transactions
-			let peer_count2 = sync.propagate_new_transactions(&mut io);
+			let peer_count2 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 			// And now the peer should have all transactions
-			let peer_count3 = sync.propagate_new_transactions(&mut io);
+			let peer_count3 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 			(peer_count2, peer_count3)
 		};
 
@@ -2347,7 +2084,7 @@ pub mod tests {
 		let queue = RwLock::new(VecDeque::new());
 		let ss = TestSnapshotService::new();
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		sync.propagate_new_transactions(&mut io);
+		SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 
 		let stats = sync.transactions_stats();
 		assert_eq!(stats.len(), 1, "Should maintain stats for single transaction.")
@@ -2377,7 +2114,7 @@ pub mod tests {
 		io.peers_info.insert(4, "Parity/v1.7.3-ABCDEFGH".to_owned());
 
 		// and new service transaction is propagated to peers
-		sync.propagate_new_transactions(&mut io);
+		SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 
 		// peer#2 && peer#4 are receiving service transaction
 		assert!(io.packets.iter().any(|p| p.packet_id == 0x02 && p.recipient == 2)); // TRANSACTIONS_PACKET
@@ -2401,7 +2138,7 @@ pub mod tests {
 		io.peers_info.insert(1, "Parity/v1.6".to_owned());
 
 		// and service + non-service transactions are propagated to peers
-		sync.propagate_new_transactions(&mut io);
+		SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
 
 		// two separate packets for peer are queued:
 		// 1) with non-service-transaction
@@ -2529,7 +2266,7 @@ pub mod tests {
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
 
 		let peers = sync.get_lagging_peers(&chain_info);
-		sync.propagate_new_hashes(&chain_info, &mut io, &peers);
+		SyncPropagator::propagate_new_hashes(&mut sync, &chain_info, &mut io, &peers);
 
 		let data = &io.packets[0].data.clone();
 		let result = sync.on_peer_new_hashes(&mut io, 0, &Rlp::new(data));
@@ -2549,7 +2286,7 @@ pub mod tests {
 		let mut io = TestIo::new(&mut client, &ss, &queue, None);
 
 		let peers = sync.get_lagging_peers(&chain_info);
-		sync.propagate_blocks(&chain_info, &mut io, &[], &peers);
+		SyncPropagator::propagate_blocks(&mut sync, &chain_info, &mut io, &[], &peers);
 
 		let data = &io.packets[0].data.clone();
 		let result = sync.on_peer_new_block(&mut io, 0, &Rlp::new(data));
